Refactor ingestion and parsing modules
